ft: add `userExceededServiceUsageLimitActivity`
