+++ apps/hash-ai-worker-ts/src/activities.ts
@@ -35,6 +35,7 @@ import {
   createEntityTypeEmbeddings,
   createPropertyTypeEmbeddings,
 } from "./activities/shared/embeddings";
+import { userExceededServiceUsageLimitActivity } from "./activities/user-exceeded-service-usage-limit-activity";
 
 export { createGraphActivities } from "./activities/graph";
 
@@ -162,6 +163,18 @@ export const createAiActivities = ({
     >,
   ) {
     return createEntitiesActivity({
+      ...params,
+      graphApiClient,
+    });
+  },
+
+  async userExceededServiceUsageLimitActivity(
+    params: Omit<
+      Parameters<typeof userExceededServiceUsageLimitActivity>[0],
+      "graphApiClient"
+    >,
+  ) {
+    return userExceededServiceUsageLimitActivity({
       ...params,
       graphApiClient,
     });
